Added error handling

+++ Application/Activities/DeleteActivity.cs
@@ -1,24 +1,31 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 
 namespace Application.Activities;
 
 public class DeleteActivity
 {
-    public class Command(Guid id) : IRequest
+    public class Command(Guid id) : IRequest<Result<Unit>>
     {
         public Guid Id { get; set; } = id;
     }
 
-    public class Handler(DataContext context) : IRequestHandler<Command>
+    public class Handler(DataContext context) : IRequestHandler<Command, Result<Unit>?>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>?> Handle(Command request, CancellationToken cancellationToken)
         {
             var activity = await context.Activities.FindAsync(request.Id);
 
+            if (activity is null) return null;
+
             context.Activities.Remove(activity);
 
-            await context.SaveChangesAsync();
+            var result = await context.SaveChangesAsync() > 0;
+
+            return result
+                ? Result<Unit>.Success(Unit.Value)
+                : Result<Unit>.Failure("Failed to delete an activity.");
         }
     }
 }
